Removed some of the files in the resources folder.
Updated:
AppStyle.java added auxiliary methods to get default styled shadows and buttons.

Corrected the value textfield in ConditionBlock.java.

Created a new Exception.
In RuleEditor.java, the method textFieldStage was updated so that if a name was in used, or the rule is incorrect, the user is warned.

In RuleFileManager.java, the method customNodeToJSON was updated so that the building blocks are checked to see if the rule is correctly made.